refactor(community): split query layer by domain
