+++ src/shared/components/Member.tsx
@@ -11,7 +11,7 @@ import { Avatar, AvatarFallback, AvatarImage } from "./ui/Avatar"
 import { UserPlus, User, Loader2 } from "lucide-react"
 import { cn } from "../lib/utils"
 import MemberDialog, { type MemberData } from "./MemberDialog"
-import { communityKeys, useCreateFriendRequest } from "../queries/community/community.queries"
+import { communityKeys, useCreateFriendRequest } from "../queries/community"
 import { useQueryClient } from "@tanstack/react-query"
 import { toast } from "sonner"
 import { useCurrentUser } from "../queries/user/user.queries"
