+++ src/shared/components/MemberDialog.tsx
@@ -4,7 +4,7 @@ import { Loader2, UserPlus } from "lucide-react"
 import { Button } from "./ui/Button"
 import { useTranslation } from "react-i18next"
 import { cn } from "../lib/utils"
-import { communityKeys, useCreateFriendRequest } from "../queries/community/community.queries"
+import { communityKeys, useCreateFriendRequest } from "../queries/community"
 import { useQueryClient } from "@tanstack/react-query"
 import { toast } from "sonner"
 import { useEffect, useCallback } from "react"
