+++ src/shared/queries/community/common/index.ts
@@ -1,0 +1,4 @@
+export interface CommunityPagination {
+  page: number
+  limit: number
+}
