+++ src/shared/queries/community/friend/index.ts
@@ -1,0 +1,5 @@
+export * from "./friend.api"
+export * from "./friend.queries"
+export * from "./friend.request"
+export * from "./friend.response"
+export * from "./friend.types"
