+++ src/shared/queries/community/member/member.request.ts
@@ -1,0 +1,3 @@
+export interface CreateMemberRequest {
+  server_id: string
+}
