+++ src/shared/queries/community/role/index.ts
@@ -1,0 +1,4 @@
+export * from "./role.api"
+export * from "./role.queries"
+export * from "./role.request"
+export * from "./role.response"
